Close #507: [`refined4s-tapir`] Reorganize type-class instances in the modules
