+++ modules/refined4s-tapir/shared/src/main/scala/refined4s/modules/tapir/derivation/types/numeric.scala
@@ -74,94 +74,94 @@ trait numeric {
 
 }
 object numeric {
-  inline given derivedNegIntSchema(using actualSchema: Schema[Int]): Schema[NegInt] = actualSchema.map(NegInt.from(_).toOption)(_.value)
+  given derivedNegIntSchema(using actualSchema: Schema[Int]): Schema[NegInt] = actualSchema.map(NegInt.from(_).toOption)(_.value)
 
-  inline given derivedNonNegIntSchema(using actualSchema: Schema[Int]): Schema[NonNegInt] =
+  given derivedNonNegIntSchema(using actualSchema: Schema[Int]): Schema[NonNegInt] =
     actualSchema.map(NonNegInt.from(_).toOption)(_.value)
 
-  inline given derivedPosIntSchema(using actualSchema: Schema[Int]): Schema[PosInt] = actualSchema.map(PosInt.from(_).toOption)(_.value)
+  given derivedPosIntSchema(using actualSchema: Schema[Int]): Schema[PosInt] = actualSchema.map(PosInt.from(_).toOption)(_.value)
 
-  inline given derivedNonPosIntSchema(using actualSchema: Schema[Int]): Schema[NonPosInt] =
+  given derivedNonPosIntSchema(using actualSchema: Schema[Int]): Schema[NonPosInt] =
     actualSchema.map(NonPosInt.from(_).toOption)(_.value)
 
-  inline given derivedNegLongSchema(using actualSchema: Schema[Long]): Schema[NegLong] = actualSchema.map(NegLong.from(_).toOption)(_.value)
+  given derivedNegLongSchema(using actualSchema: Schema[Long]): Schema[NegLong] = actualSchema.map(NegLong.from(_).toOption)(_.value)
 
-  inline given derivedNonNegLongSchema(using actualSchema: Schema[Long]): Schema[NonNegLong] =
+  given derivedNonNegLongSchema(using actualSchema: Schema[Long]): Schema[NonNegLong] =
     actualSchema.map(NonNegLong.from(_).toOption)(_.value)
 
-  inline given derivedPosLongSchema(using actualSchema: Schema[Long]): Schema[PosLong] = actualSchema.map(PosLong.from(_).toOption)(_.value)
+  given derivedPosLongSchema(using actualSchema: Schema[Long]): Schema[PosLong] = actualSchema.map(PosLong.from(_).toOption)(_.value)
 
-  inline given derivedNonPosLongSchema(using actualSchema: Schema[Long]): Schema[NonPosLong] =
+  given derivedNonPosLongSchema(using actualSchema: Schema[Long]): Schema[NonPosLong] =
     actualSchema.map(NonPosLong.from(_).toOption)(_.value)
 
-  inline given derivedNegShortSchema(using actualSchema: Schema[Short]): Schema[NegShort] =
+  given derivedNegShortSchema(using actualSchema: Schema[Short]): Schema[NegShort] =
     actualSchema.map(NegShort.from(_).toOption)(_.value)
 
-  inline given derivedNonNegShortSchema(using actualSchema: Schema[Short]): Schema[NonNegShort] =
+  given derivedNonNegShortSchema(using actualSchema: Schema[Short]): Schema[NonNegShort] =
     actualSchema.map(NonNegShort.from(_).toOption)(_.value)
 
-  inline given derivedPosShortSchema(using actualSchema: Schema[Short]): Schema[PosShort] =
+  given derivedPosShortSchema(using actualSchema: Schema[Short]): Schema[PosShort] =
     actualSchema.map(PosShort.from(_).toOption)(_.value)
 
-  inline given derivedNonPosShortSchema(using actualSchema: Schema[Short]): Schema[NonPosShort] =
+  given derivedNonPosShortSchema(using actualSchema: Schema[Short]): Schema[NonPosShort] =
     actualSchema.map(NonPosShort.from(_).toOption)(_.value)
 
-  inline given derivedNegByteSchema(using actualSchema: Schema[Byte]): Schema[NegByte] = actualSchema.map(NegByte.from(_).toOption)(_.value)
+  given derivedNegByteSchema(using actualSchema: Schema[Byte]): Schema[NegByte] = actualSchema.map(NegByte.from(_).toOption)(_.value)
 
-  inline given derivedNonNegByteSchema(using actualSchema: Schema[Byte]): Schema[NonNegByte] =
+  given derivedNonNegByteSchema(using actualSchema: Schema[Byte]): Schema[NonNegByte] =
     actualSchema.map(NonNegByte.from(_).toOption)(_.value)
 
-  inline given derivedPosByteSchema(using actualSchema: Schema[Byte]): Schema[PosByte] = actualSchema.map(PosByte.from(_).toOption)(_.value)
+  given derivedPosByteSchema(using actualSchema: Schema[Byte]): Schema[PosByte] = actualSchema.map(PosByte.from(_).toOption)(_.value)
 
-  inline given derivedNonPosByteSchema(using actualSchema: Schema[Byte]): Schema[NonPosByte] =
+  given derivedNonPosByteSchema(using actualSchema: Schema[Byte]): Schema[NonPosByte] =
     actualSchema.map(NonPosByte.from(_).toOption)(_.value)
 
-  inline given derivedNegFloatSchema(using actualSchema: Schema[Float]): Schema[NegFloat] =
+  given derivedNegFloatSchema(using actualSchema: Schema[Float]): Schema[NegFloat] =
     actualSchema.map(NegFloat.from(_).toOption)(_.value)
 
-  inline given derivedNonNegFloatSchema(using actualSchema: Schema[Float]): Schema[NonNegFloat] =
+  given derivedNonNegFloatSchema(using actualSchema: Schema[Float]): Schema[NonNegFloat] =
     actualSchema.map(NonNegFloat.from(_).toOption)(_.value)
 
-  inline given derivedPosFloatSchema(using actualSchema: Schema[Float]): Schema[PosFloat] =
+  given derivedPosFloatSchema(using actualSchema: Schema[Float]): Schema[PosFloat] =
     actualSchema.map(PosFloat.from(_).toOption)(_.value)
 
-  inline given derivedNonPosFloatSchema(using actualSchema: Schema[Float]): Schema[NonPosFloat] =
+  given derivedNonPosFloatSchema(using actualSchema: Schema[Float]): Schema[NonPosFloat] =
     actualSchema.map(NonPosFloat.from(_).toOption)(_.value)
 
-  inline given derivedNegDoubleSchema(using actualSchema: Schema[Double]): Schema[NegDouble] =
+  given derivedNegDoubleSchema(using actualSchema: Schema[Double]): Schema[NegDouble] =
     actualSchema.map(NegDouble.from(_).toOption)(_.value)
 
-  inline given derivedNonNegDoubleSchema(using actualSchema: Schema[Double]): Schema[NonNegDouble] =
+  given derivedNonNegDoubleSchema(using actualSchema: Schema[Double]): Schema[NonNegDouble] =
     actualSchema.map(NonNegDouble.from(_).toOption)(_.value)
 
-  inline given derivedPosDoubleSchema(using actualSchema: Schema[Double]): Schema[PosDouble] =
+  given derivedPosDoubleSchema(using actualSchema: Schema[Double]): Schema[PosDouble] =
     actualSchema.map(PosDouble.from(_).toOption)(_.value)
 
-  inline given derivedNonPosDoubleSchema(using actualSchema: Schema[Double]): Schema[NonPosDouble] =
+  given derivedNonPosDoubleSchema(using actualSchema: Schema[Double]): Schema[NonPosDouble] =
     actualSchema.map(NonPosDouble.from(_).toOption)(_.value)
 
-  inline given derivedNegBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NegBigInt] =
+  given derivedNegBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NegBigInt] =
     actualSchema.map(NegBigInt.from(_).toOption)(_.value)
 
-  inline given derivedNonNegBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NonNegBigInt] =
+  given derivedNonNegBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NonNegBigInt] =
     actualSchema.map(NonNegBigInt.from(_).toOption)(_.value)
 
-  inline given derivedPosBigIntSchema(using actualSchema: Schema[BigInt]): Schema[PosBigInt] =
+  given derivedPosBigIntSchema(using actualSchema: Schema[BigInt]): Schema[PosBigInt] =
     actualSchema.map(PosBigInt.from(_).toOption)(_.value)
 
-  inline given derivedNonPosBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NonPosBigInt] =
+  given derivedNonPosBigIntSchema(using actualSchema: Schema[BigInt]): Schema[NonPosBigInt] =
     actualSchema.map(NonPosBigInt.from(_).toOption)(_.value)
 
-  inline given derivedNegBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NegBigDecimal] =
+  given derivedNegBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NegBigDecimal] =
     actualSchema.map(NegBigDecimal.from(_).toOption)(_.value)
 
-  inline given derivedNonNegBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NonNegBigDecimal] =
+  given derivedNonNegBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NonNegBigDecimal] =
     actualSchema.map(NonNegBigDecimal.from(_).toOption)(_.value)
 
-  inline given derivedPosBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[PosBigDecimal] =
+  given derivedPosBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[PosBigDecimal] =
     actualSchema.map(PosBigDecimal.from(_).toOption)(_.value)
 
-  inline given derivedNonPosBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NonPosBigDecimal] =
+  given derivedNonPosBigDecimalSchema(using actualSchema: Schema[BigDecimal]): Schema[NonPosBigDecimal] =
     actualSchema.map(NonPosBigDecimal.from(_).toOption)(_.value)
 
 }
